initial presenter
