+++ app/build.gradle
@@ -56,6 +56,5 @@ dependencies {
     implementation 'com.google.android.libraries.places:places:2.1.0'
 
     // rxjava
-    implementation 'io.reactivex.rxjava2:rxandroid:2.1.1'
-
+    implementation 'io.reactivex.rxjava2:rxandroid:2.0.1'
 }
